refactor(admin-ui): split RealmDetailPage.tsx into per-tab settings forms
RealmDetailPage.tsx had grown to 1345 lines with one giant shared form
covering all 9 realm settings tabs, a single mega-mutation, and every
tab's data queries running unconditionally regardless of which tab was
active. Split each tab into a self-contained component (General,
Tokens, Security, Events, Theme, Locale) following the pattern already
used by EmailProviderForm/SmsProviderForm/MagicLinkSettingsForm: each
owns its form state, its own partial-body PATCH mutation, and its own
success/error banner.

Also:
- Added the missing WebAuthn/rate-limit/impersonation/locale fields to
  the Realm type (confirmed non-null in the Prisma schema) so the new
  Security/Locale forms type-check without `as any` casts.
- Hoisted the duplicated formatDuration() into a shared util; the copy
  in SmsProviderForm lacked a day-rollover branch, so durations >= 1 day
  now render as "Xd" there too, matching the other forms.
