+++ admin-ui/src/pages/realms/RealmDetailPage.tsx
@@ -1,26 +1,17 @@
-import { useState, useEffect, type FormEvent } from 'react';
-import { useParams, useNavigate, Link } from 'react-router';
+import { useState } from 'react';
+import { useParams, useNavigate } from 'react-router';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { getRealmByName, updateRealm, deleteRealm, exportRealm, getThemes } from '../../api/realms';
-import type { ThemeInfo } from '../../api/realms';
-import { getUsers } from '../../api/users';
-import { getClients } from '../../api/clients';
-import { getRealmRoles } from '../../api/roles';
-import { getGroups } from '../../api/groups';
-import { getClientScopes } from '../../api/clientScopes';
-import { getRealmSessions } from '../../api/sessions';
-import { getIdentityProviders } from '../../api/identityProviders';
+import { getRealmByName, deleteRealm, exportRealm } from '../../api/realms';
 import ConfirmDialog from '../../components/ConfirmDialog';
 import MagicLinkSettingsForm from '../../components/magic-link/MagicLinkSettingsForm';
 import EmailProviderForm from '../../components/email/EmailProviderForm';
 import SmsProviderForm from '../../components/sms/SmsProviderForm';
-
-function formatDuration(seconds: number): string {
-  if (seconds < 60) return `${seconds}s`;
-  if (seconds < 3600) return `${Math.floor(seconds / 60)}m`;
-  if (seconds < 86400) return `${Math.floor(seconds / 3600)}h`;
-  return `${Math.floor(seconds / 86400)}d`;
-}
+import GeneralSettingsForm from '../../components/realms/GeneralSettingsForm';
+import TokenSettingsForm from '../../components/realms/TokenSettingsForm';
+import SecuritySettingsForm from '../../components/realms/SecuritySettingsForm';
+import EventSettingsForm from '../../components/realms/EventSettingsForm';
+import ThemeSettingsForm from '../../components/realms/ThemeSettingsForm';
+import LocaleSettingsForm from '../../components/realms/LocaleSettingsForm';
 
 export default function RealmDetailPage() {
   const { name } = useParams<{ name: string }>();
@@ -35,199 +26,6 @@ export default function RealmDetailPage() {
     enabled: !!name,
   });
 
-  // Sub-resource lists are loaded lazily per tab so opening a realm doesn't fire
-  // a thundering herd of parallel list calls on mount. The overview counts below
-  // are only shown on the General tab; the theme list only on the Theme tab.
-  const overviewEnabled = !!name && !!realm && activeTab === 'general';
-
-  const { data: users } = useQuery({
-    queryKey: ['users', name],
-    queryFn: () => getUsers(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: clients } = useQuery({
-    queryKey: ['clients', name],
-    queryFn: () => getClients(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: roles } = useQuery({
-    queryKey: ['roles', name],
-    queryFn: () => getRealmRoles(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: groups } = useQuery({
-    queryKey: ['groups', name],
-    queryFn: () => getGroups(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: clientScopes } = useQuery({
-    queryKey: ['clientScopes', name],
-    queryFn: () => getClientScopes(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: sessions } = useQuery({
-    queryKey: ['sessions', name],
-    queryFn: () => getRealmSessions(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: identityProviders } = useQuery({
-    queryKey: ['identity-providers', name],
-    queryFn: () => getIdentityProviders(name!),
-    enabled: overviewEnabled,
-  });
-
-  const { data: themes } = useQuery({
-    queryKey: ['themes'],
-    queryFn: () => getThemes(),
-    enabled: activeTab === 'theme',
-  });
-
-  const [form, setForm] = useState({
-    displayName: '',
-    enabled: true,
-    registrationAllowed: true,
-    accessTokenLifespan: 300,
-    refreshTokenLifespan: 1800,
-    offlineTokenLifespan: 2592000,
-    // Security - Password Policy
-    passwordMinLength: 8,
-    passwordRequireUppercase: false,
-    passwordRequireLowercase: false,
-    passwordRequireDigits: false,
-    passwordRequireSpecialChars: false,
-    passwordHistoryCount: 0,
-    passwordMaxAgeDays: 0,
-    // Security - Brute Force Protection
-    bruteForceEnabled: false,
-    maxLoginFailures: 5,
-    lockoutDuration: 300,
-    failureResetTime: 600,
-    permanentLockoutAfter: 0,
-    // Security - MFA
-    mfaRequired: false,
-    // WebAuthn
-    webAuthnEnabled: false,
-    webAuthnRpName: '',
-    webAuthnRpId: '',
-    webAuthnUserVerificationRequired: false,
-    // Rate Limiting
-    rateLimitEnabled: false,
-    clientRateLimitPerMinute: 100,
-    clientRateLimitPerHour: 1000,
-    userRateLimitPerMinute: 30,
-    userRateLimitPerHour: 300,
-    ipRateLimitPerMinute: 60,
-    ipRateLimitPerHour: 600,
-    // Impersonation
-    impersonationEnabled: false,
-    impersonationMaxDuration: 1800,
-    // Risk Thresholds
-    riskThresholdStepUp: 50,
-    riskThresholdBlock: 80,
-    // Session Limits
-    maxSessionsPerUser: 0,
-    // Adaptive Authentication
-    adaptiveAuthEnabled: false,
-    // Locale
-    defaultLocale: 'en',
-    supportedLocales: [] as string[],
-    // Events
-    eventsEnabled: false,
-    eventsExpiration: 604800,
-    adminEventsEnabled: false,
-    loginEventRetentionDays: 30,
-    adminEventRetentionDays: 90,
-    deletionGracePeriodDays: 14,
-    // Theme
-    themeName: 'idenplane',
-    loginTheme: 'idenplane',
-    accountTheme: 'idenplane',
-    emailTheme: 'idenplane',
-  });
-
-  // Seed the editable form from fetched data when the loaded realm changes.
-  // Adjusting state during render (vs. an effect) avoids an extra render pass.
-  const [seededRealm, setSeededRealm] = useState(realm);
-  if (realm && realm !== seededRealm) {
-    setSeededRealm(realm);
-    setForm({
-        displayName: realm.displayName ?? '',
-        enabled: realm.enabled,
-        registrationAllowed: realm.registrationAllowed ?? true,
-        accessTokenLifespan: realm.accessTokenLifespan,
-        refreshTokenLifespan: realm.refreshTokenLifespan,
-        offlineTokenLifespan: realm.offlineTokenLifespan ?? 2592000,
-        // Security - Password Policy
-        passwordMinLength: realm.passwordMinLength ?? 8,
-        passwordRequireUppercase: realm.passwordRequireUppercase ?? false,
-        passwordRequireLowercase: realm.passwordRequireLowercase ?? false,
-        passwordRequireDigits: realm.passwordRequireDigits ?? false,
-        passwordRequireSpecialChars: realm.passwordRequireSpecialChars ?? false,
-        passwordHistoryCount: realm.passwordHistoryCount ?? 0,
-        passwordMaxAgeDays: realm.passwordMaxAgeDays ?? 0,
-        // Security - Brute Force Protection
-        bruteForceEnabled: realm.bruteForceEnabled ?? false,
-        maxLoginFailures: realm.maxLoginFailures ?? 5,
-        lockoutDuration: realm.lockoutDuration ?? 300,
-        failureResetTime: realm.failureResetTime ?? 600,
-        permanentLockoutAfter: realm.permanentLockoutAfter ?? 0,
-        // Security - MFA
-        mfaRequired: realm.mfaRequired ?? false,
-        // WebAuthn
-        webAuthnEnabled: (realm as any).webAuthnEnabled ?? false,
-        webAuthnRpName: (realm as any).webAuthnRpName ?? '',
-        webAuthnRpId: (realm as any).webAuthnRpId ?? '',
-        webAuthnUserVerificationRequired: (realm as any).webAuthnUserVerificationRequired ?? false,
-        // Rate Limiting
-        rateLimitEnabled: (realm as any).rateLimitEnabled ?? false,
-        clientRateLimitPerMinute: (realm as any).clientRateLimitPerMinute ?? 100,
-        clientRateLimitPerHour: (realm as any).clientRateLimitPerHour ?? 1000,
-        userRateLimitPerMinute: (realm as any).userRateLimitPerMinute ?? 30,
-        userRateLimitPerHour: (realm as any).userRateLimitPerHour ?? 300,
-        ipRateLimitPerMinute: (realm as any).ipRateLimitPerMinute ?? 60,
-        ipRateLimitPerHour: (realm as any).ipRateLimitPerHour ?? 600,
-        // Impersonation
-        impersonationEnabled: (realm as any).impersonationEnabled ?? false,
-        impersonationMaxDuration: (realm as any).impersonationMaxDuration ?? 1800,
-        // Risk Thresholds
-        riskThresholdStepUp: (realm as any).riskThresholdStepUp ?? 50,
-        riskThresholdBlock: (realm as any).riskThresholdBlock ?? 80,
-        // Session Limits
-        maxSessionsPerUser: (realm as any).maxSessionsPerUser ?? 0,
-        // Adaptive Authentication
-        adaptiveAuthEnabled: (realm as any).adaptiveAuthEnabled ?? false,
-        // Locale
-        defaultLocale: (realm as any).defaultLocale ?? 'en',
-        supportedLocales: (realm as any).supportedLocales ?? [],
-        // Events
-        eventsEnabled: realm.eventsEnabled ?? false,
-        eventsExpiration: realm.eventsExpiration ?? 604800,
-        adminEventsEnabled: realm.adminEventsEnabled ?? false,
-        loginEventRetentionDays: (realm as any).loginEventRetentionDays ?? 30,
-        adminEventRetentionDays: (realm as any).adminEventRetentionDays ?? 90,
-        deletionGracePeriodDays: (realm as any).deletionGracePeriodDays ?? 14,
-        // Theme
-        themeName: realm.themeName ?? 'idenplane',
-        loginTheme: realm.loginTheme ?? 'idenplane',
-        accountTheme: realm.accountTheme ?? 'idenplane',
-        emailTheme: realm.emailTheme ?? 'idenplane',
-    });
-  }
-
-  const updateMutation = useMutation({
-    mutationFn: () => updateRealm(name!, form),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['realm', name] });
-      queryClient.invalidateQueries({ queryKey: ['realms'] });
-    },
-  });
-
   const deleteMutation = useMutation({
     mutationFn: () => deleteRealm(name!),
     onSuccess: () => {
@@ -235,17 +33,6 @@ export default function RealmDetailPage() {
       navigate('/console/realms');
     },
   });
-
-  // Clear toast messages when switching tabs
-  useEffect(() => {
-    updateMutation.reset();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeTab]);
-
-  function handleSubmit(e: FormEvent) {
-    e.preventDefault();
-    updateMutation.mutate();
-  }
 
   if (isLoading) {
     return (
@@ -273,16 +60,6 @@ export default function RealmDetailPage() {
     { key: 'theme' as const, label: 'Theme' },
     { key: 'magic-link' as const, label: 'Magic Link' },
     { key: 'locale' as const, label: 'Locale' },
-  ];
-
-  const quickLinks = [
-    { to: `/console/realms/${name}/users`, label: 'Users', count: users?.total },
-    { to: `/console/realms/${name}/clients`, label: 'Clients', count: clients?.length },
-    { to: `/console/realms/${name}/roles`, label: 'Roles', count: roles?.length },
-    { to: `/console/realms/${name}/groups`, label: 'Groups', count: groups?.length },
-    { to: `/console/realms/${name}/client-scopes`, label: 'Client Scopes', count: clientScopes?.length },
-    { to: `/console/realms/${name}/sessions`, label: 'Sessions', count: sessions?.length },
-    { to: `/console/realms/${name}/identity-providers`, label: 'Identity Providers', count: identityProviders?.length },
   ];
 
   return (
@@ -339,999 +116,15 @@ export default function RealmDetailPage() {
         </nav>
       </div>
 
-      {/* General Tab */}
-      {activeTab === 'general' && (
-        <div className="space-y-8">
-          {/* Quick links */}
-          <div className="grid gap-4 sm:grid-cols-3">
-            {quickLinks.map((link) => (
-              <Link
-                key={link.to}
-                to={link.to}
-                className="flex items-center justify-between rounded-lg border border-line bg-surface p-4 shadow-sm hover:shadow-md transition-shadow"
-              >
-                <div>
-                  <p className="text-sm font-medium text-subtle">{link.label}</p>
-                  <p className="text-2xl font-bold text-fg">
-                    {link.count !== undefined ? link.count : '-'}
-                  </p>
-                </div>
-                <svg className="h-5 w-5 text-subtle" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-                </svg>
-              </Link>
-            ))}
-          </div>
-
-          {/* General settings form */}
-          <form onSubmit={handleSubmit} className="space-y-6 rounded-lg border border-line bg-surface p-6 shadow-sm">
-            <h2 className="text-lg font-semibold text-fg">General Settings</h2>
-
-            <div>
-              <label htmlFor="realm-name" className="mb-1.5 block text-sm font-medium text-muted">Name</label>
-              <input
-                id="realm-name"
-                type="text"
-                value={realm.name}
-                disabled
-                className="w-full rounded-md border border-line bg-sunken px-3 py-2 text-sm text-subtle"
-              />
-              <p className="mt-1 text-xs text-subtle">Realm name cannot be changed</p>
-            </div>
-
-            <div>
-              <label htmlFor="realm-display-name" className="mb-1.5 block text-sm font-medium text-muted">Display Name</label>
-              <input
-                id="realm-display-name"
-                type="text"
-                value={form.displayName}
-                onChange={(e) => setForm({ ...form, displayName: e.target.value })}
-                className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-            </div>
-
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="enabled"
-                checked={form.enabled}
-                onChange={(e) => setForm({ ...form, enabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="enabled" className="text-sm font-medium text-muted">
-                Enabled
-              </label>
-            </div>
-
-            <div>
-              <div className="flex items-center gap-2">
-                <input
-                  type="checkbox"
-                  id="registrationAllowed"
-                  checked={form.registrationAllowed}
-                  onChange={(e) => setForm({ ...form, registrationAllowed: e.target.checked })}
-                  className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-                />
-                <label htmlFor="registrationAllowed" className="text-sm font-medium text-muted">
-                  User Registration
-                </label>
-              </div>
-              <p className="mt-1 ml-6 text-xs text-subtle">When disabled, users cannot self-register. Only admins can create accounts.</p>
-            </div>
-
-            {updateMutation.isSuccess && (
-              <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-                Realm updated successfully.
-              </div>
-            )}
-            {updateMutation.isError && (
-              <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-                Failed to update realm.
-              </div>
-            )}
-
-            <div className="flex justify-end border-t border-line pt-4">
-              <button
-                type="submit"
-                disabled={updateMutation.isPending}
-                className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-              >
-                {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-              </button>
-            </div>
-          </form>
-        </div>
-      )}
-
-      {/* Tokens Tab */}
-      {activeTab === 'tokens' && (
-        <form onSubmit={handleSubmit} className="space-y-6 rounded-lg border border-line bg-surface p-6 shadow-sm">
-          <h2 className="text-lg font-semibold text-fg">Token Lifespans</h2>
-          <p className="text-sm text-subtle">
-            Configure how long access and refresh tokens remain valid.
-          </p>
-
-          <div className="space-y-6">
-            <div>
-              <label htmlFor="access-token-lifespan" className="mb-1.5 block text-sm font-medium text-muted">
-                Access Token Lifespan
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  id="access-token-lifespan"
-                  type="number"
-                  min={1}
-                  value={form.accessTokenLifespan}
-                  onChange={(e) =>
-                    setForm({ ...form, accessTokenLifespan: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="text-sm text-subtle">seconds</span>
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.accessTokenLifespan)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                How long an access token is valid before it must be refreshed.
-              </p>
-            </div>
-
-            <div>
-              <label htmlFor="refresh-token-lifespan" className="mb-1.5 block text-sm font-medium text-muted">
-                Refresh Token Lifespan
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  id="refresh-token-lifespan"
-                  type="number"
-                  min={1}
-                  value={form.refreshTokenLifespan}
-                  onChange={(e) =>
-                    setForm({ ...form, refreshTokenLifespan: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="text-sm text-subtle">seconds</span>
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.refreshTokenLifespan)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                How long a refresh token is valid. This also controls session duration.
-              </p>
-            </div>
-
-            <div>
-              <label htmlFor="offline-token-lifespan" className="mb-1.5 block text-sm font-medium text-muted">
-                Offline Token Lifespan
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  id="offline-token-lifespan"
-                  type="number"
-                  min={60}
-                  value={form.offlineTokenLifespan}
-                  onChange={(e) =>
-                    setForm({ ...form, offlineTokenLifespan: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="text-sm text-subtle">seconds</span>
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.offlineTokenLifespan)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                How long offline refresh tokens are valid. These survive session logout. Default: 30 days.
-              </p>
-            </div>
-          </div>
-
-          {updateMutation.isSuccess && (
-            <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-              Token settings updated successfully.
-            </div>
-          )}
-          {updateMutation.isError && (
-            <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-              Failed to update token settings.
-            </div>
-          )}
-
-          <div className="flex justify-end border-t border-line pt-4">
-            <button
-              type="submit"
-              disabled={updateMutation.isPending}
-              className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-            >
-              {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-            </button>
-          </div>
-        </form>
-      )}
-
-      {/* Email Tab */}
+      {activeTab === 'general' && <GeneralSettingsForm realm={realm} />}
+      {activeTab === 'tokens' && <TokenSettingsForm realm={realm} />}
       {activeTab === 'email' && <EmailProviderForm realm={realm} />}
-
-      {/* SMS Tab */}
       {activeTab === 'sms' && <SmsProviderForm realm={realm} />}
-
-      {/* Security Tab */}
-      {activeTab === 'security' && (
-        <form onSubmit={handleSubmit} className="space-y-8 rounded-lg border border-line bg-surface p-6 shadow-sm">
-          {/* Password Policy */}
-          <div className="space-y-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Password Policy</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Define password complexity and rotation requirements for users in this realm.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Minimum Password Length
-              </label>
-              <input
-                type="number"
-                min={1}
-                value={form.passwordMinLength}
-                onChange={(e) =>
-                  setForm({ ...form, passwordMinLength: Number(e.target.value) })
-                }
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-            </div>
-
-            <div className="space-y-3">
-              <div className="flex items-center gap-2">
-                <input
-                  type="checkbox"
-                  id="passwordRequireUppercase"
-                  checked={form.passwordRequireUppercase}
-                  onChange={(e) =>
-                    setForm({ ...form, passwordRequireUppercase: e.target.checked })
-                  }
-                  className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-                />
-                <label htmlFor="passwordRequireUppercase" className="text-sm font-medium text-muted">
-                  Require uppercase letters
-                </label>
-              </div>
-
-              <div className="flex items-center gap-2">
-                <input
-                  type="checkbox"
-                  id="passwordRequireLowercase"
-                  checked={form.passwordRequireLowercase}
-                  onChange={(e) =>
-                    setForm({ ...form, passwordRequireLowercase: e.target.checked })
-                  }
-                  className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-                />
-                <label htmlFor="passwordRequireLowercase" className="text-sm font-medium text-muted">
-                  Require lowercase letters
-                </label>
-              </div>
-
-              <div className="flex items-center gap-2">
-                <input
-                  type="checkbox"
-                  id="passwordRequireDigits"
-                  checked={form.passwordRequireDigits}
-                  onChange={(e) =>
-                    setForm({ ...form, passwordRequireDigits: e.target.checked })
-                  }
-                  className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-                />
-                <label htmlFor="passwordRequireDigits" className="text-sm font-medium text-muted">
-                  Require digits
-                </label>
-              </div>
-
-              <div className="flex items-center gap-2">
-                <input
-                  type="checkbox"
-                  id="passwordRequireSpecialChars"
-                  checked={form.passwordRequireSpecialChars}
-                  onChange={(e) =>
-                    setForm({ ...form, passwordRequireSpecialChars: e.target.checked })
-                  }
-                  className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-                />
-                <label htmlFor="passwordRequireSpecialChars" className="text-sm font-medium text-muted">
-                  Require special characters
-                </label>
-              </div>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Password History Count
-              </label>
-              <input
-                type="number"
-                min={0}
-                value={form.passwordHistoryCount}
-                onChange={(e) =>
-                  setForm({ ...form, passwordHistoryCount: Number(e.target.value) })
-                }
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                Number of previous passwords to remember. Users cannot reuse these.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Password Max Age (days)
-              </label>
-              <input
-                type="number"
-                min={0}
-                value={form.passwordMaxAgeDays}
-                onChange={(e) =>
-                  setForm({ ...form, passwordMaxAgeDays: Number(e.target.value) })
-                }
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                0 = no expiry. Users will be required to change their password after this many days.
-              </p>
-            </div>
-          </div>
-
-          {/* Brute Force Protection */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Brute Force Protection</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Protect user accounts from brute force login attacks.
-              </p>
-            </div>
-
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="bruteForceEnabled"
-                checked={form.bruteForceEnabled}
-                onChange={(e) =>
-                  setForm({ ...form, bruteForceEnabled: e.target.checked })
-                }
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="bruteForceEnabled" className="text-sm font-medium text-muted">
-                Enable brute force protection
-              </label>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Max Login Failures
-              </label>
-              <input
-                type="number"
-                min={1}
-                value={form.maxLoginFailures}
-                onChange={(e) =>
-                  setForm({ ...form, maxLoginFailures: Number(e.target.value) })
-                }
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                Number of consecutive login failures before the account is locked.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Lockout Duration (seconds)
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  type="number"
-                  min={1}
-                  value={form.lockoutDuration}
-                  onChange={(e) =>
-                    setForm({ ...form, lockoutDuration: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.lockoutDuration)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                How long an account is locked after exceeding max failures.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Failure Reset Time (seconds)
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  type="number"
-                  min={1}
-                  value={form.failureResetTime}
-                  onChange={(e) =>
-                    setForm({ ...form, failureResetTime: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.failureResetTime)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                Time after which the failure counter is reset if no new failures occur.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Permanent Lockout After
-              </label>
-              <input
-                type="number"
-                min={0}
-                value={form.permanentLockoutAfter}
-                onChange={(e) =>
-                  setForm({ ...form, permanentLockoutAfter: Number(e.target.value) })
-                }
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                0 = disabled. Number of temporary lockouts before the account is permanently locked.
-              </p>
-            </div>
-          </div>
-
-          {/* MFA */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Multi-Factor Authentication</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Configure MFA requirements for this realm.
-              </p>
-            </div>
-
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="mfaRequired"
-                checked={form.mfaRequired}
-                onChange={(e) =>
-                  setForm({ ...form, mfaRequired: e.target.checked })
-                }
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="mfaRequired" className="text-sm font-medium text-muted">
-                Require all users to set up TOTP
-              </label>
-            </div>
-          </div>
-
-          {/* WebAuthn */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">WebAuthn / Passkeys</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Allow users to authenticate with hardware security keys and passkeys.
-              </p>
-            </div>
-            <div className="flex items-center gap-2">
-              <input type="checkbox" id="webAuthnEnabled" checked={form.webAuthnEnabled}
-                onChange={(e) => setForm({ ...form, webAuthnEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent" />
-              <label htmlFor="webAuthnEnabled" className="text-sm font-medium text-muted">Enable WebAuthn</label>
-            </div>
-            {form.webAuthnEnabled && (
-              <div className="space-y-4 pl-6">
-                <div>
-                  <label className="mb-1.5 block text-sm font-medium text-muted">Relying Party Name</label>
-                  <input type="text" value={form.webAuthnRpName}
-                    onChange={(e) => setForm({ ...form, webAuthnRpName: e.target.value })}
-                    placeholder="My App"
-                    className="w-full max-w-md rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  <p className="mt-1 text-xs text-subtle">Display name shown to users during passkey registration</p>
-                </div>
-                <div>
-                  <label className="mb-1.5 block text-sm font-medium text-muted">Relying Party ID (domain)</label>
-                  <input type="text" value={form.webAuthnRpId}
-                    onChange={(e) => setForm({ ...form, webAuthnRpId: e.target.value })}
-                    placeholder="example.com"
-                    className="w-full max-w-md rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  <p className="mt-1 text-xs text-subtle">Must match your domain. Credentials registered here only work on this domain.</p>
-                </div>
-                <div className="flex items-center gap-2">
-                  <input type="checkbox" id="webAuthnUserVerificationRequired" checked={form.webAuthnUserVerificationRequired}
-                    onChange={(e) => setForm({ ...form, webAuthnUserVerificationRequired: e.target.checked })}
-                    className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent" />
-                  <label htmlFor="webAuthnUserVerificationRequired" className="text-sm font-medium text-muted">Require user verification (biometric/PIN)</label>
-                </div>
-              </div>
-            )}
-          </div>
-
-          {/* Rate Limiting */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Rate Limiting</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Throttle authentication requests to prevent abuse.
-              </p>
-            </div>
-            <div className="flex items-center gap-2">
-              <input type="checkbox" id="rateLimitEnabled" checked={form.rateLimitEnabled}
-                onChange={(e) => setForm({ ...form, rateLimitEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent" />
-              <label htmlFor="rateLimitEnabled" className="text-sm font-medium text-muted">Enable rate limiting</label>
-            </div>
-            {form.rateLimitEnabled && (
-              <div className="space-y-4 pl-6">
-                <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">Client req/min</label>
-                    <input type="number" min={1} value={form.clientRateLimitPerMinute}
-                      onChange={(e) => setForm({ ...form, clientRateLimitPerMinute: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">Client req/hour</label>
-                    <input type="number" min={1} value={form.clientRateLimitPerHour}
-                      onChange={(e) => setForm({ ...form, clientRateLimitPerHour: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">User req/min</label>
-                    <input type="number" min={1} value={form.userRateLimitPerMinute}
-                      onChange={(e) => setForm({ ...form, userRateLimitPerMinute: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">User req/hour</label>
-                    <input type="number" min={1} value={form.userRateLimitPerHour}
-                      onChange={(e) => setForm({ ...form, userRateLimitPerHour: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">IP req/min</label>
-                    <input type="number" min={1} value={form.ipRateLimitPerMinute}
-                      onChange={(e) => setForm({ ...form, ipRateLimitPerMinute: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                  <div>
-                    <label className="mb-1.5 block text-sm font-medium text-muted">IP req/hour</label>
-                    <input type="number" min={1} value={form.ipRateLimitPerHour}
-                      onChange={(e) => setForm({ ...form, ipRateLimitPerHour: Number(e.target.value) })}
-                      className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-
-          {/* Impersonation */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Impersonation</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Allow admins to impersonate users for debugging and support.
-              </p>
-            </div>
-            <div className="flex items-center gap-2">
-              <input type="checkbox" id="impersonationEnabled" checked={form.impersonationEnabled}
-                onChange={(e) => setForm({ ...form, impersonationEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent" />
-              <label htmlFor="impersonationEnabled" className="text-sm font-medium text-muted">Allow admin impersonation</label>
-            </div>
-            {form.impersonationEnabled && (
-              <div className="pl-6">
-                <label className="mb-1.5 block text-sm font-medium text-muted">Max impersonation duration (seconds)</label>
-                <div className="flex items-center gap-3">
-                  <input type="number" min={60} value={form.impersonationMaxDuration}
-                    onChange={(e) => setForm({ ...form, impersonationMaxDuration: Number(e.target.value) })}
-                    className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                  <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                    {formatDuration(form.impersonationMaxDuration)}
-                  </span>
-                </div>
-                <p className="mt-1 text-xs text-subtle">Default: 30 minutes. Impersonation tokens expire after this duration.</p>
-              </div>
-            )}
-          </div>
-
-          {/* Risk Thresholds */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Risk Thresholds</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Score boundaries (0–100) that trigger step-up authentication or session block.
-              </p>
-            </div>
-            <div className="grid grid-cols-1 gap-6 sm:grid-cols-2">
-              <div>
-                <label className="mb-1.5 block text-sm font-medium text-muted">Step-up threshold</label>
-                <input type="number" min={0} max={100} value={form.riskThresholdStepUp}
-                  onChange={(e) => setForm({ ...form, riskThresholdStepUp: Number(e.target.value) })}
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                <p className="mt-1 text-xs text-subtle">Risk score ≥ this value triggers MFA step-up. Default: 50.</p>
-              </div>
-              <div>
-                <label className="mb-1.5 block text-sm font-medium text-muted">Block threshold</label>
-                <input type="number" min={0} max={100} value={form.riskThresholdBlock}
-                  onChange={(e) => setForm({ ...form, riskThresholdBlock: Number(e.target.value) })}
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                <p className="mt-1 text-xs text-subtle">Risk score ≥ this value blocks the session entirely. Default: 80.</p>
-              </div>
-            </div>
-          </div>
-
-          {/* Session Limits */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Session Limits</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Limit the number of concurrent active sessions per user.
-              </p>
-            </div>
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">Max sessions per user</label>
-              <input type="number" min={0} value={form.maxSessionsPerUser}
-                onChange={(e) => setForm({ ...form, maxSessionsPerUser: Number(e.target.value) })}
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-              <p className="mt-1 text-xs text-subtle">
-                Maximum concurrent active sessions per user. When the limit is reached the oldest session is evicted.
-                Set to 0 for unlimited.
-              </p>
-            </div>
-          </div>
-
-          {/* Adaptive Authentication */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Adaptive Authentication</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Enable AI-powered risk scoring to dynamically adjust authentication requirements.
-              </p>
-            </div>
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="adaptiveAuthEnabled"
-                checked={form.adaptiveAuthEnabled}
-                onChange={(e) => setForm({ ...form, adaptiveAuthEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="adaptiveAuthEnabled" className="text-sm font-medium text-muted">
-                Enable adaptive authentication
-              </label>
-            </div>
-            <p className="ml-6 -mt-4 text-xs text-subtle">
-              When enabled, each login attempt is scored for risk. High-risk logins trigger step-up MFA or are
-              blocked based on the thresholds above.
-            </p>
-          </div>
-
-          {updateMutation.isSuccess && (
-            <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-              Security settings updated successfully.
-            </div>
-          )}
-          {updateMutation.isError && (
-            <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-              Failed to update security settings.
-            </div>
-          )}
-
-          <div className="flex justify-end border-t border-line pt-4">
-            <button
-              type="submit"
-              disabled={updateMutation.isPending}
-              className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-            >
-              {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-            </button>
-          </div>
-        </form>
-      )}
-
-      {/* Events Tab */}
-      {activeTab === 'events' && (
-        <form onSubmit={handleSubmit} className="space-y-8 rounded-lg border border-line bg-surface p-6 shadow-sm">
-          <div>
-            <h2 className="text-lg font-semibold text-fg">Event Configuration</h2>
-            <p className="mt-1 text-sm text-subtle">
-              Control whether login and admin events are recorded for this realm.
-            </p>
-          </div>
-
-          <div className="space-y-6">
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="eventsEnabled"
-                checked={form.eventsEnabled}
-                onChange={(e) => setForm({ ...form, eventsEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="eventsEnabled" className="text-sm font-medium text-muted">
-                Enable login events
-              </label>
-            </div>
-            <p className="ml-6 -mt-4 text-xs text-subtle">
-              Records login, logout, token refresh, and authentication error events.
-            </p>
-
-            <div className="flex items-center gap-2">
-              <input
-                type="checkbox"
-                id="adminEventsEnabled"
-                checked={form.adminEventsEnabled}
-                onChange={(e) => setForm({ ...form, adminEventsEnabled: e.target.checked })}
-                className="h-4 w-4 rounded border-line-strong text-accent focus:ring-accent"
-              />
-              <label htmlFor="adminEventsEnabled" className="text-sm font-medium text-muted">
-                Enable admin events
-              </label>
-            </div>
-            <p className="ml-6 -mt-4 text-xs text-subtle">
-              Records create, update, and delete operations performed through the admin API.
-            </p>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">
-                Events Expiration (seconds)
-              </label>
-              <div className="flex items-center gap-3">
-                <input
-                  type="number"
-                  min={60}
-                  value={form.eventsExpiration}
-                  onChange={(e) =>
-                    setForm({ ...form, eventsExpiration: Number(e.target.value) })
-                  }
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                />
-                <span className="text-sm text-subtle">seconds</span>
-                <span className="rounded-full bg-sunken px-2.5 py-0.5 text-xs font-medium text-muted">
-                  {formatDuration(form.eventsExpiration)}
-                </span>
-              </div>
-              <p className="mt-1 text-xs text-subtle">
-                How long events are kept before automatic cleanup. Default: 7 days (604800s).
-              </p>
-            </div>
-          </div>
-
-          {/* Event Retention */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">Event Retention</h2>
-              <p className="mt-1 text-sm text-subtle">
-                How many days event records are retained in the database before being purged.
-              </p>
-            </div>
-            <div className="grid grid-cols-1 gap-6 sm:grid-cols-2">
-              <div>
-                <label className="mb-1.5 block text-sm font-medium text-muted">Login event retention (days)</label>
-                <input type="number" min={1} value={form.loginEventRetentionDays}
-                  onChange={(e) => setForm({ ...form, loginEventRetentionDays: Number(e.target.value) })}
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                <p className="mt-1 text-xs text-subtle">Days to keep login/auth event records. Default: 30.</p>
-              </div>
-              <div>
-                <label className="mb-1.5 block text-sm font-medium text-muted">Admin event retention (days)</label>
-                <input type="number" min={1} value={form.adminEventRetentionDays}
-                  onChange={(e) => setForm({ ...form, adminEventRetentionDays: Number(e.target.value) })}
-                  className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-                <p className="mt-1 text-xs text-subtle">Days to keep admin audit event records. Default: 90.</p>
-              </div>
-            </div>
-          </div>
-
-          {/* User Lifecycle */}
-          <div className="space-y-6 border-t border-line pt-6">
-            <div>
-              <h2 className="text-lg font-semibold text-fg">User Lifecycle</h2>
-              <p className="mt-1 text-sm text-subtle">
-                Configure how long deleted user accounts are retained before permanent removal.
-              </p>
-            </div>
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">Deletion grace period (days)</label>
-              <input type="number" min={0} value={form.deletionGracePeriodDays}
-                onChange={(e) => setForm({ ...form, deletionGracePeriodDays: Number(e.target.value) })}
-                className="w-40 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none" />
-              <p className="mt-1 text-xs text-subtle">
-                Days before a soft-deleted account is permanently purged. Set to 0 to disable soft-deletion. Default: 14.
-              </p>
-            </div>
-          </div>
-
-          {updateMutation.isSuccess && (
-            <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-              Event settings updated successfully.
-            </div>
-          )}
-          {updateMutation.isError && (
-            <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-              Failed to update event settings.
-            </div>
-          )}
-
-          <div className="flex justify-end border-t border-line pt-4">
-            <button
-              type="submit"
-              disabled={updateMutation.isPending}
-              className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-            >
-              {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-            </button>
-          </div>
-        </form>
-      )}
-
-      {/* Theme Tab */}
-      {activeTab === 'theme' && (
-        <form onSubmit={handleSubmit} className="space-y-8 rounded-lg border border-line bg-surface p-6 shadow-sm">
-          <div>
-            <h2 className="text-lg font-semibold text-fg">Theme Settings</h2>
-            <p className="mt-1 text-sm text-subtle">
-              Assign themes per page type and customize colors for this realm.
-            </p>
-          </div>
-
-          {/* Per-Type Theme Selectors */}
-          {themes && themes.length > 0 && (
-            <div className="space-y-4">
-              <h3 className="text-sm font-semibold uppercase tracking-wider text-subtle">Theme Assignment</h3>
-              <div className="grid grid-cols-1 gap-4 sm:grid-cols-3">
-                <div>
-                  <label className="mb-1.5 block text-sm font-medium text-muted">Login Theme</label>
-                  <select
-                    value={form.loginTheme}
-                    onChange={(e) => setForm({ ...form, loginTheme: e.target.value, themeName: e.target.value })}
-                    className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                  >
-                    {themes.map((t: ThemeInfo) => (
-                      <option key={t.name} value={t.name}>{t.displayName}</option>
-                    ))}
-                  </select>
-                  <p className="mt-1 text-xs text-subtle">Applied to login, register, consent, and other auth pages</p>
-                </div>
-                <div>
-                  <label className="mb-1.5 block text-sm font-medium text-muted">Account Theme</label>
-                  <select
-                    value={form.accountTheme}
-                    onChange={(e) => setForm({ ...form, accountTheme: e.target.value })}
-                    className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                  >
-                    {themes.map((t: ThemeInfo) => (
-                      <option key={t.name} value={t.name}>{t.displayName}</option>
-                    ))}
-                  </select>
-                  <p className="mt-1 text-xs text-subtle">Applied to account management and TOTP setup pages</p>
-                </div>
-                <div>
-                  <label className="mb-1.5 block text-sm font-medium text-muted">Email Theme</label>
-                  <select
-                    value={form.emailTheme}
-                    onChange={(e) => setForm({ ...form, emailTheme: e.target.value })}
-                    className="w-full rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-                  >
-                    {themes.map((t: ThemeInfo) => (
-                      <option key={t.name} value={t.name}>{t.displayName}</option>
-                    ))}
-                  </select>
-                  <p className="mt-1 text-xs text-subtle">Applied to verification and password reset emails</p>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {updateMutation.isSuccess && (
-            <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-              Theme settings updated successfully.
-            </div>
-          )}
-          {updateMutation.isError && (
-            <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-              Failed to update theme settings.
-            </div>
-          )}
-
-          <div className="flex justify-end border-t border-line pt-4">
-            <button
-              type="submit"
-              disabled={updateMutation.isPending}
-              className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-            >
-              {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-            </button>
-          </div>
-        </form>
-      )}
-
-      {/* Magic Link Tab */}
-      {activeTab === 'magic-link' && realm && (
-        <MagicLinkSettingsForm realm={realm} />
-      )}
-
-      {/* Locale Tab */}
-      {activeTab === 'locale' && (
-        <form onSubmit={handleSubmit} className="space-y-8 rounded-lg border border-line bg-surface p-6 shadow-sm">
-          <div>
-            <h2 className="text-lg font-semibold text-fg">Locale &amp; Internationalization</h2>
-            <p className="mt-1 text-sm text-subtle">
-              Configure the default language and supported locales for this realm.
-            </p>
-          </div>
-
-          <div className="space-y-6">
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">Default Locale</label>
-              <input
-                type="text"
-                value={form.defaultLocale}
-                onChange={(e) => setForm({ ...form, defaultLocale: e.target.value })}
-                placeholder="en"
-                className="w-48 rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                BCP 47 language tag (e.g. en, en-US, fr, de, ar). Used when no user preference is set.
-              </p>
-            </div>
-
-            <div>
-              <label className="mb-1.5 block text-sm font-medium text-muted">Supported Locales</label>
-              <input
-                type="text"
-                value={form.supportedLocales.join(', ')}
-                onChange={(e) =>
-                  setForm({
-                    ...form,
-                    supportedLocales: e.target.value
-                      .split(',')
-                      .map((s) => s.trim())
-                      .filter(Boolean),
-                  })
-                }
-                placeholder="en, fr, de, ar"
-                className="w-full max-w-md rounded-md border border-line-strong px-3 py-2 text-sm shadow-sm focus:border-accent focus:ring-1 focus:ring-accent focus:outline-none"
-              />
-              <p className="mt-1 text-xs text-subtle">
-                Comma-separated list of supported locale codes. Users can switch between these in the account UI.
-              </p>
-            </div>
-          </div>
-
-          {updateMutation.isSuccess && (
-            <div className="rounded-md bg-success-soft p-3 text-sm text-success-fg">
-              Locale settings updated successfully.
-            </div>
-          )}
-          {updateMutation.isError && (
-            <div className="rounded-md bg-danger-soft p-3 text-sm text-danger-fg">
-              Failed to update locale settings.
-            </div>
-          )}
-
-          <div className="flex justify-end border-t border-line pt-4">
-            <button
-              type="submit"
-              disabled={updateMutation.isPending}
-              className="rounded-md bg-accent px-4 py-2 text-sm font-medium text-white hover:bg-accent-hover disabled:opacity-50"
-            >
-              {updateMutation.isPending ? 'Saving...' : 'Save Changes'}
-            </button>
-          </div>
-        </form>
-      )}
+      {activeTab === 'security' && <SecuritySettingsForm realm={realm} />}
+      {activeTab === 'events' && <EventSettingsForm realm={realm} />}
+      {activeTab === 'theme' && <ThemeSettingsForm realm={realm} />}
+      {activeTab === 'magic-link' && <MagicLinkSettingsForm realm={realm} />}
+      {activeTab === 'locale' && <LocaleSettingsForm realm={realm} />}
 
       <ConfirmDialog
         isOpen={showDelete}
