+++ admin-ui/src/test/mocks/data.ts
@@ -33,10 +33,25 @@ export function makeRealm(overrides: Partial<Realm> = {}): Realm {
     permanentLockoutAfter: 0,
     registrationAllowed: false,
     mfaRequired: false,
+    webAuthnEnabled: false,
+    webAuthnRpName: null,
+    webAuthnRpId: null,
+    webAuthnUserVerificationRequired: false,
     offlineTokenLifespan: 2592000,
     eventsEnabled: false,
     eventsExpiration: 0,
     adminEventsEnabled: false,
+    rateLimitEnabled: false,
+    clientRateLimitPerMinute: 60,
+    clientRateLimitPerHour: 1000,
+    userRateLimitPerMinute: 30,
+    userRateLimitPerHour: 500,
+    ipRateLimitPerMinute: 20,
+    ipRateLimitPerHour: 200,
+    impersonationEnabled: false,
+    impersonationMaxDuration: 1800,
+    defaultLocale: 'en',
+    supportedLocales: ['en'],
     themeName: 'default',
     theme: null,
     loginTheme: 'default',
